Fix state transition when file loading fails
There's an annoying discrepancy in the protocol -- if you reject a
segment request from the main boot routine, it doesn't return the ack
sequence (0x10 0x06). But it does return that when you reject a
request while loading a program from the menu, getting a time packet,
etc.

Since I'd rather have a clean response under the latter circumstances,
and it doesn't really make any difference (the unexpected situation
will result in a reset and resync anyway), go back to sending
unauthorized.