Formatting: convert tabs to spaces
Tabs should never be used for indentation in source
code. Unfortunately, the skeleton of this app was generated by
preact-cli, and the default template used tabs. I didn't see this
because I'm using Visual Studio Code and it detects the indentation
and switches its settings to match.

This abomination has now been corrected.