Refactor state machine organization
What was known as `adaptor` is now `protocol'.
What was known as `serial` is now `adaptor`.

The main entry point is now (the new) `adaptor`. This will be a state
machine that manages getting a port (either serial or socket) and
invoking a protocol machine.

Code in the application will think of this as the adaptor, and not
care about how its internals are coupled. Communication with the app
will be handled through reading and writing the machine context,
and sending events.